// Frontend/SUBD-client/src/app/login/interfaces/sign-up-credentials.ts
export interface ISignUpCredentials {
    userEmail: string;
    password: string;
    dni: string;
    name: string;
    lastName: string;
    province: string;
    municipality: string;
    city: string;
    postalCode: string;
    emergencyContact: string;
    healthInsurance: string;
    insurancePlan: string;
}